pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "network-toolkit"
dynamic = ["version"]
description = "Network automation toolkit for MikroTik RouterOS and beyond"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
keywords = ["network", "automation", "mikrotik", "routeros", "scrapli"]
authors = [
    { name = "Network Team", email = "network@company.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
]
dependencies = [
    "scrapli[paramiko]>=2023.7.30",
    "scrapli[community]>=2023.7.30",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "typer[all]",
    "rich>=13.0.0",
    "asyncssh>=2.21.0",
    "scrapli-community>=2025.1.30",
    "libtmux>=0.46.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "detect-secrets>=1.4.0",
]
tmux = [
    "libtmux>=0.21.0",
]

[project.urls]
Documentation = "https://github.com/your-org/network-toolkit#readme"
Issues = "https://github.com/your-org/network-toolkit/issues"
Source = "https://github.com/your-org/network-toolkit"

[project.scripts]
netkit = "network_toolkit.cli:app"

[tool.hatch.version]
path = "src/network_toolkit/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
    "pytest-asyncio",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov = [
    "test-cov",
    "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
    "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/network_toolkit tests}"

[tool.coverage.run]
source_pkgs = ["network_toolkit", "tests"]
branch = true
parallel = true
omit = [
    "src/network_toolkit/__about__.py",
]

[tool.coverage.paths]
network_toolkit = ["src/network_toolkit", "*/network-toolkit/src/network_toolkit"]
tests = ["tests", "*/network-toolkit/tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# Package build configuration (hatchling)
# - Ensure non-Python assets needed at runtime are included in wheels
# - Provide a source distribution that contains helpful docs and scripts
[tool.hatch.build.targets.wheel]
packages = ["src/network_toolkit"]
include = [
    # Typing marker for consumers
    "src/network_toolkit/py.typed",
    # Built-in sequence YAMLs needed at runtime
    "src/network_toolkit/builtin_sequences/**/*.yml",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/**",
    "README.md",
    "LICENSE*",
    "docs/**",
    # Shell completion helpers and example defaults (useful for users)
    "bash_completion_*",
    "fish_completion_*",
    "zsh_completion_*",
    "default_configs/**",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Allow boolean positional values in function calls, like `dict.get(... True)`
    "FBT003",
    # Ignore checks for possible passwords
    "S105", "S106", "S107",
    # Ignore complexity
    "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
    # Don't touch unused imports
    "F401",
]

[tool.ruff.isort]
known-first-party = ["network_toolkit"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
]
